Add functionalities, refactor more proffesional

+++ include/fibonacci.h
@@ -2,7 +2,7 @@
 #define FIBONACCI_H
 #include <stdint.h>
 
-uint64_t fibonacci_calculator(uint64_t n);
+uint32_t fibonacci_calculator(int n);
 
 #endif
